fix: use client for creating app upload
